Created the Searoom navigation command which generates new searooms as needed.
